requirements.txt
requests~=2.27.1
coincurve~=17.0.0
pysha3~=1.0.2